Make VM functions use lapi:: for consistency
