+++ include/blt/lapi_vm.hh
@@ -1,0 +1,11 @@
+
+#pragma once
+
+namespace blt {
+    namespace lapi {
+        namespace vm {
+            void base_open(lua_state*);
+        }
+    }
+}
+
