--- include/blt/vmexport.hh
@@ -1,9 +1,0 @@
-
-#pragma once
-
-namespace blt {
-	namespace vmexport {
-		void base_open(lua_state*);
-	}
-}
-
