+++ src/hook.cc
@@ -27,7 +27,7 @@ extern "C" {
 #include <blt/lapi.hh>
 #include <blt/lapi_systemfs.hh>
 #include <blt/lapi_version.hh>
-#include <blt/vmexport.hh>
+#include <blt/lapi_vm.hh>
 
 
 /**
@@ -195,7 +195,7 @@ namespace blt {
             };
             luaL_openlib(state, "SystemFS", lib_SystemFS, 0);
 
-            vmexport::base_open(state);
+            lapi::vm::base_open(state);
         }
